Handle hero template variations.
